Struct google-siteverification1::WebResourceMethodsBuilder

pub struct WebResourceMethodsBuilder<'a, C, NC, A> where C: 'a, NC: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on webResource resources. It is not used directly, but through the SiteVerification hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate "google-siteverification1" as siteverification1;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use siteverification1::SiteVerification;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = SiteVerification::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `delete(...)`, `get(...)`, `get_token(...)`, `insert(...)`, `list(...)`, `patch(...)` and `update(...)`
// to build up your call.
let rb = hub.web_resource();

Methods

impl<'a, C, NC, A> WebResourceMethodsBuilder<'a, C, NC, A>

fn get(&self, id: &str) -> WebResourceGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get the most current data for a website or domain.

fn patch(&self, request: &SiteVerificationWebResourceResource, id: &str) -> WebResourcePatchMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modify the list of owners for your website or domain. This method supports patch semantics.

fn list(&self) -> WebResourceListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get the list of your verified websites and domains.

fn delete(&self, id: &str) -> WebResourceDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Relinquish ownership of a website or domain.

fn get_token(&self, request: &SiteVerificationWebResourceGettokenRequest) -> WebResourceGetTokenMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Get a verification token for placing on a website or domain.

fn update(&self, request: &SiteVerificationWebResourceResource, id: &str) -> WebResourceUpdateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modify the list of owners for your website or domain.

fn insert(&self, request: &SiteVerificationWebResourceResource, verification_method: &str) -> WebResourceInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Attempt verification of a website or domain.